+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.6
+## 0.0.7
 
 * First version
 ---
@@ -12,3 +12,16 @@ Mode
 ```
 
 - Adding a method for open a map using Google map
+- Adding method to send message to whatsapp
+
+```dart
+ MaterialButton(
+    color: Colors.green[700],
+    textColor: Colors.white,
+    onPressed: () async {
+        await EasyLauncher.sendToWhatsApp(
+             phone: "+93700000000", message: "hi");
+        },
+    child: const Text("Send to whatsapp"),
+),
+```